+++ tests/convergence_tests/check_for_meshes.py
@@ -6,14 +6,14 @@ try:
     import requests
     import clint
 except ImportError:
-    sys.exit("Regression tests require `requests` and `clint`, please install using pip or conda")
+    sys.exit("Convergence tests require `requests` and `clint`, please install using pip or conda")
 
 def download_zip_with_progress_bar(url):
     r = requests.get(url, stream=True)
     path = url.rsplit('/', 1)[-1]
     with open(path, 'wb') as f:
         total_length = int(r.headers.get('content-length'))
-        for chunk in clint.textui.progress.bar(r.iter_content(chunk_size=1024), expected_size=(total_length/1024) + 1): 
+        for chunk in clint.textui.progress.bar(r.iter_content(chunk_size=1024), expected_size=(total_length/1024) + 1):
             if chunk:
                 f.write(chunk)
                 f.flush()
@@ -31,7 +31,7 @@ def unzip(meshzip):
 def check_mesh():
     #check if there's a geometry folder present in the directory
     if not os.path.isdir('geometry'):
-        dl_check = input('The meshes for regression tests don\'t appear to '
+        dl_check = input('The meshes for convergence tests don\'t appear to '
                          'be loaded. Would you like to download them from '
                              'Zenodo? (~10MB) (y/n): ')
         if dl_check == 'y':