+++ tests/convergence_tests/run_convergence_tests.py
@@ -1,15 +1,17 @@
-import subprocess
 import os
+import time
+import subprocess
+import datetime
 
 from check_for_meshes import check_mesh
 
-#tests to run
+# tests to run
 tests = ['lysozyme.py', 'molecule_dirichlet.py', 'molecule_neumann.py',
          'sphere_dirichlet.py', 'sphere_molecule_single.py',
          'sphere_molecule_stern.py', 'sphere_neumann.py',
          'twosphere_dirichlet.py', 'twosphere_neumann.py', 'two_molecules.py']
 
-#specify CUDA device to use
+# specify CUDA device to use
 CUDA_DEVICE = '0'
 
 ENV = os.environ.copy()
@@ -17,11 +19,12 @@ ENV['CUDA_DEVICE'] = CUDA_DEVICE
 
 check_mesh()
 
+tic = time.time()
+
 for test in tests:
     subprocess.call(['python', '{}'.format(test)])
 
-#remove test result cache
-try:
-    os.remove('tests')
-except IOError:
-    pass
+toc = time.time()
+
+print("Total runtime for convergence tests: ")
+print(str(datetime.timedelta(seconds=(toc - tic))))