+++ tests/convergence_tests/sphere_molecule_stern.py
@@ -1,9 +1,8 @@
 import numpy
-import pickle
 
 from pygbe.util import an_solution
-from convergence import (scanOutput, run_convergence, picklesave, pickleload,
-                        report_results, mesh)
+from convergence import (run_convergence, picklesave, pickleload,
+                         report_results, mesh)
 
 
 def main():
@@ -15,7 +14,7 @@ def main():
 
     problem_folder = 'input_files'
 
-    #molecule_stern
+    # molecule_stern
     param = 'sphere_fine.param'
     test_name = 'molecule_stern'
     if test_name not in test_outputs.keys():
@@ -25,7 +24,7 @@ def main():
 
     picklesave(test_outputs)
 
-    #load results for analysis
+    # load results for analysis
     Esolv, Esurf, Ecoul = test_outputs['molecule_stern'][2:5]
     Time = test_outputs['molecule_stern'][-1]
     N = test_outputs['molecule_stern'][0]
@@ -44,7 +43,8 @@ def main():
                    Esolv,
                    analytical,
                    total_time,
-                   energy_type='Total')
+                   energy_type='Total',
+                   test_name='sphere molecule stern')
 
 
 if __name__ == "__main__":