+++ tests/test_sphere.py
@@ -19,15 +19,25 @@ from pygbe.main import main
                                  'E_coul_kcal',
                                  'E_coul_kJ',
                                  'E_solv_kcal'])
-def test_PGB_mut_sensor(key):
+def test_sphere(key):
     results = get_results()
 
     with open('sphere.pickle', 'rb') as f:
         base_results = pickle.load(f)
 
-    assert base_results[key] == results[key]
+    if base_results[key] > 0:
+        assert abs(base_results[key] - results[key]) / base_results[key] < 1e-12
+    else:
+        assert base_results[key] == results[key]
 
-@functools.lru_cache(5)
+def test_sphere_iterations():
+    results = get_results()
+    with open('sphere.pickle', 'rb') as f:
+        base_results = pickle.load(f)
+
+    assert base_results['iterations'] == results['iterations']
+
+@functools.lru_cache(6)
 def get_results():
     print('Generating results for sphere example...')
     if os.getcwd().rsplit('/', 1)[1] == 'tests':